Adjust flour proportions to account for stickiness
The dough has been coming out very sticky and is hard for the food processor to handle. This should also increase the gluten due to the decrease of rye flour (with very little, and poor quality, gluten) and increasing the bread flour. Hopefully, this makes for an easier to work with dough all around.

Additionally, the rye and wheat tastes could be pretty strong without proper proofing. They're both reduced a bit in this recipe, so we'll see how that goes. I checked the published Mozza pizza dough recipe, and it only has about 0.2% rye to bread flour. This change brings our proportions from almost 10.5% rye/bread flour to almost 6.5%. Still far more than Mozza, but quite a bit less for this recipe. We'll see how it goes!